requirements.txt
asgiref==3.3.1
click==7.1.2
cycler==0.10.0
Django==3.1.7
django-crispy-forms==1.11.2
imutils==0.5.4
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
xlwt==1.3.0
